bluetooth: split the ring buffer library out
