+++ ring.h
@@ -1,0 +1,47 @@
+/*
+ * This file is part of the MicroPython for Monocle:
+ *      https://github.com/Itsbrilliantlabs/monocle-micropython
+ *
+ * Authored by: Raj Nakarja - Silicon Witchery AB
+ * Authored by: Josuah Demangeon - Panoramix Labs
+ *
+ * ISC Licence
+ *
+ * Copyright © 2022 Brilliant Labs Inc.
+ *
+ * Permission to use, copy, modify, and/or distribute this software for any
+ * purpose with or without fee is hereby granted, provided that the above
+ * copyright notice and this permission notice appear in all copies.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES WITH
+ * REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF MERCHANTABILITY
+ * AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR ANY SPECIAL, DIRECT,
+ * INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES WHATSOEVER RESULTING FROM
+ * LOSS OF USE, DATA OR PROFITS, WHETHER IN AN ACTION OF CONTRACT, NEGLIGENCE OR
+ * OTHER TORTIOUS ACTION, ARISING OUT OF OR IN CONNECTION WITH THE USE OR
+ * PERFORMANCE OF THIS SOFTWARE.
+ */
+
+#pragma once
+
+#include <stdint.h>
+#include <stddef.h>
+#include <stdbool.h>
+
+// Buffer sizes for REPL ring buffers; +45 allows a bytearray to be printed in one go.
+#define RING_BUFFER_LENGTH (1024 + 45)
+
+/**
+ * @brief Ring buffers for the repl rx and tx data which goes over BLE.
+ */
+typedef struct
+{
+    uint8_t buffer[RING_BUFFER_LENGTH];
+    uint16_t head;
+    uint16_t tail;
+} ring_buf_t;
+
+extern bool ring_full(ring_buf_t const *ring);
+extern bool ring_empty(ring_buf_t const *ring);
+extern void ring_push(ring_buf_t *ring, uint8_t byte);
+extern uint8_t ring_pop(ring_buf_t *ring);
